Refactor: Update imports and module exports for ES Modules
