+++ models/Cliente.js
@@ -7,4 +7,4 @@ class Cliente {
   }
 }
 
-module.exports = Cliente;
+export default Cliente;
