+++ models/Evento.js
@@ -10,5 +10,5 @@ class Evento {
   }
 }
 
-module.exports = Evento;
+export default Evento;
